pay_api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pay-zhifubao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "alipay-sdk-python>=3.7.1098",
    "cryptography>=42.0.0",
    "fastapi>=0.115.0",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
# 在 pay_api 目录执行: uv run serve
serve = "pay_api.main:main"

[tool.setuptools]
packages = ["pay_api"]

[tool.setuptools.package-dir]
pay_api = "."

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[tool.basedpyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
